new bug resolving

+++ GUI/exp_delete.py
@@ -128,7 +128,9 @@ class Delete_Exp(QDialog):
         try:
             conn= sqlite3.connect(db_path)
             cursor= conn.cursor()
-            cursor.execute("delete from products  where name=? and quantity=?",(name,quantity))
+            cursor.execute("select id from users limit 1")
+            id_user= cursor.fetchone()[0]
+            cursor.execute("delete from products  where name=? and quantity=? and user_id=?",(name,quantity,id_user))
             conn.commit()
             MessageBox(text="محصول موفقانه حذف شد",title="موفقانه",type="info").show()
             self.name_line.clear()
